Fixed .conf generation bug and moved plugin images directory
Duplicate entries were being created in generated .conf file due to a
>> where there should only have been >.

Also changed where plugin stores the images that are copied into the
emhttp directory when plugin is installed. They were in a general
directory, but there was a possible clash with icons from plugins by
overbyrn (on which this plugin is based).